pyproject.toml
[project]
name = "my-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.2.0",
    "faker>=36.1.0",
    "fastapi>=0.115.8",
    "httpx>=0.28.1",
    "markdown>=3.7",
    "pillow>=11.1.0",
    "pytesseract>=0.3.13",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.9.6",
    "sentence-transformers>=3.4.1",
    "uvicorn>=0.34.0",
]
